+++ sdk/resourcemanager/dns/armdns/client_factory.go
@@ -24,7 +24,7 @@ type ClientFactory struct {
 
 // NewClientFactory creates a new instance of ClientFactory with the specified values.
 // The parameter values will be propagated to any client created from this factory.
-//   - subscriptionID - Specifies the Azure subscription ID, which uniquely identifies the Microsoft Azure subscription.
+//   - subscriptionID - The ID of the target subscription.
 //   - credential - used to authorize requests. Usually a credential from azidentity.
 //   - options - pass nil to accept the default values.
 func NewClientFactory(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ClientFactory, error) {
@@ -36,6 +36,11 @@ func NewClientFactory(subscriptionID string, credential azcore.TokenCredential, 
 		subscriptionID: subscriptionID, credential: credential,
 		options: options.Clone(),
 	}, nil
+}
+
+func (c *ClientFactory) NewDnssecConfigsClient() *DnssecConfigsClient {
+	subClient, _ := NewDnssecConfigsClient(c.subscriptionID, c.credential, c.options)
+	return subClient
 }
 
 func (c *ClientFactory) NewRecordSetsClient() *RecordSetsClient {