+++ sdk/resourcemanager/dns/armdns/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Release History
 
+## 1.2.0-beta.1 (2023-05-17)
+### Features Added
+
+- New value `RecordTypeDS`, `RecordTypeNAPTR`, `RecordTypeTLSA` added to enum type `RecordType`
+- New enum type `CreatedByType` with values `CreatedByTypeApplication`, `CreatedByTypeKey`, `CreatedByTypeManagedIdentity`, `CreatedByTypeUser`
+- New function `*ClientFactory.NewDnssecConfigsClient() *DnssecConfigsClient`
+- New function `NewDnssecConfigsClient(string, azcore.TokenCredential, *arm.ClientOptions) (*DnssecConfigsClient, error)`
+- New function `*DnssecConfigsClient.BeginCreateOrUpdate(context.Context, string, string, *DnssecConfigsClientBeginCreateOrUpdateOptions) (*runtime.Poller[DnssecConfigsClientCreateOrUpdateResponse], error)`
+- New function `*DnssecConfigsClient.BeginDelete(context.Context, string, string, *DnssecConfigsClientBeginDeleteOptions) (*runtime.Poller[DnssecConfigsClientDeleteResponse], error)`
+- New function `*DnssecConfigsClient.Get(context.Context, string, string, *DnssecConfigsClientGetOptions) (DnssecConfigsClientGetResponse, error)`
+- New function `*DnssecConfigsClient.NewListByDNSZonePager(string, string, *DnssecConfigsClientListByDNSZoneOptions) *runtime.Pager[DnssecConfigsClientListByDNSZoneResponse]`
+- New function `timeRFC3339.MarshalText() ([]byte, error)`
+- New function `*timeRFC3339.Parse(string) error`
+- New function `*timeRFC3339.UnmarshalText([]byte) error`
+- New struct `DelegationSignerInfo`
+- New struct `Digest`
+- New struct `DnssecConfig`
+- New struct `DnssecConfigListResult`
+- New struct `DnssecProperties`
+- New struct `DsRecord`
+- New struct `NaptrRecord`
+- New struct `SigningKey`
+- New struct `SystemData`
+- New struct `TlsaRecord`
+- New field `DsRecords`, `NaptrRecords`, `TlsaRecords` in struct `RecordSetProperties`
+- New field `SystemData` in struct `Zone`
+- New field `SigningKeys` in struct `ZoneProperties`
+
+
 ## 1.1.0 (2023-03-28)
 ### Features Added